requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# CLIP
open-clip-torch>=2.20.0

# DINOv2 dependency
timm>=0.9.0

# Data + evaluation
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.8.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
